[PATCH] udev init script
I integrated udev with Fedora Core.  The main piece is simply building
/udev on boot, since we don't have an initramfs yet. We should also   
clear out /udev on shutdown, for /udev directories mounted on persistent
media.

The attached script goes in /etc/init.d

Then do "chkconfig --add udev"

And the rest is handled automatically.  I made it for Fedora but it will
probably work, with little change, on any Linux system.

Right now it only does sysfs-based discovery of block and tty devices,
since those are the only types of devices I have on my system.  There is
a TODO in the script where we would add the other device types.