[PATCH] Writing udev rules doc update
Here's an update to the documentation. It makes a few minor corrections and
adds info about multiple-symlinks.

It also seems that the patch I sent on April 27th (patching v0.53 to 0.54) was
not applied, so this patch includes that update too, which was also just some
small corrections plus added info on renaming network devices.