[PATCH] respect prefix= setting in built udev.conf (updated)
Here is a revised version of the patch. Again, it modifies the Makefile
to respect the prefix= setting when putting paths to
/etc/udev/{rules.s,permissions.d} into the built /etc/udev/udev.conf
file. It also changes the Makefile to create this file at "make" time,
not "make install" time. This allows for udevdir to be specified at
"make" time (thus putting the correct path into udev.conf), but not
specified at "make install" time (thus allowing the installation to
proceed without trying to use the wrong directory).

 Submitted By: Kevin P. Fleming <kpfleming@linuxfromscratch.org>
 Date: 2004-09-16
 Initial Package Version: 032
 Origin: David Jensen
 Description: correct udev's Makefile and template config file
 to respect the "prefix=" setting supplied when it is built; also
 build etc/udev/udev.conf at "make" time, not "make install" time