[PATCH] update the man pages and correct Usage: hints
Add UDEV_LOG to the man udev man page. Remove mention of specific
variables from the udevd/udevsend man page as we changed to pass
the whole environment.

Correct printed Usage: of udevtest and udevinfo.

Init the config in udevtest earlier to accept input with and without
the sysfs mount point.