[PATCH] Dialout group fix for capi devices in the gentoo rules file
currently it is not possible to use the capi devices for users even
when the user is in the dialout group. Please apply the attached patch
and fix the issue.

There is also a gentoo bug about this:
http://bugs.gentoo.org/show_bug.cgi?id=92445

diff -ur udev-056.orig/etc/udev/gentoo/udev.rules udev-056/etc/udev/gentoo/udev.rules