run_program: close pipe fd's which are connected to child process
When udev_log="debug", the children of udev inherit the file
descriptors of the pipes used to capture output. If they are not
properly closed then udev will not exit until all programs started
in the background by the children are terminated or have closed the
fds themselves, and this may cause deadlocks with udevsettle.