hid2hci: narrow matches to real HCI devices
The previous rules just checked bInterfaceProtocol but not the actual device
and interface class. This caused the hci rules to be applied for Dell USB hubs
and attached input devices like keyboards and mouses as well, breaking them
completely.

Tighten the match to also check device and interface class/subclass.

https://launchpad.net/bugs/392144